Modifications necessary for compilation on Fedora Core 38 to complete.
For consideration:
1.
Compiling also succeeds if I remove the libomp directory from the following line:
include_directories(faiss/ libomp/omp/build/runtime/src)

2.
Do we really want to build libomp ourselves?

3.
If we do:
then, do we really want to link to OpenBLAS?

Given that the current CMake config worked for Hannes but not for me,
it seems that Hannes also links against a BLAS library on Apple.
Maybe it is always present on Apple OS?