Add mutability flag on indexes.
Indexes loaded from disk lose all information we need to determine if an index is trained.
Because of this we cannot guarantee consistent training with these indexes.
If the index was still untrained, we do allow adding of new data.
In the future it may be nice to provide a function to manually mark them as mutable.